añadidos enlaces para navegacion facil, ordenacion de listados, paginacion de listados y filtrado de listas de favoritos y de pedidos por usuario

+++ DSS/database/migrations/2021_03_02_153026_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             $table->timestamps();
             $table->bigInteger('user_id')->unsigned()->index()->nullable();
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('totalPrice');
+            $table->float('totalPrice');
             $table->string('paymentMethod')->nullable();
         });
     }
